# backend-go/openapi.yaml
openapi: 3.0.3
info:
  title: Cancer Detection Backend API
  version: 1.0.0
  description: |
    Gateway-facing backend API.
    Internal microservice communication may use gRPC, but external clients should use these HTTP endpoints.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/audits:
    get:
      tags: [Audit]
      summary: List audit logs (DOCTOR/ADMIN)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: action
          schema: { type: string }
        - in: query
          name: entityType
          schema: { type: string }
        - in: query
          name: entityId
          schema: { type: string }
        - in: query
          name: result
          schema: { type: string }
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items: { type: object, additionalProperties: true }
                  nextCursor: { type: string, nullable: true }
  /api/v1/analytics/admin-overview:
    get:
      tags: [Analytics]
      summary: Admin overview
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/ops/readiness:
    get:
      tags: [Ops]
      summary: Ops readiness
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /api/v1/ops/dashboard:
    get:
      tags: [Ops]
      summary: Ops dashboard
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /api/v1/ops/evidence:
    get:
      tags: [Ops]
      summary: List clinical evidence
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: { type: object, additionalProperties: true }
    post:
      tags: [Ops]
      summary: Create clinical evidence (ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceCreateRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/ops/incidents:
    get:
      tags: [Ops]
      summary: List incidents
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [OPEN, ACKNOWLEDGED, RESOLVED] }
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: { type: object, additionalProperties: true }
    post:
      tags: [Ops]
      summary: Create incident (ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentCreateRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/ops/incidents/{id}/transition:
    post:
      tags: [Ops]
      summary: Transition incident (ADMIN)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [ACKNOWLEDGE, RESOLVE, REOPEN]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema: { type: string }
    LimitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          nullable: true
...
    EvidenceCreateRequest:
      type: object
      required:
        - runName
        - modelVersion
        - datasetName
        - sampleCount
        - positiveCount
        - siteCount
        - auroc
        - sensitivity
        - specificity
      properties:
        runName: { type: string }
        modelVersion: { type: string }
        datasetName: { type: string }
        datasetVersion: { type: string }
        sampleCount: { type: integer }
        positiveCount: { type: integer }
        siteCount: { type: integer }
        auroc: { type: number, format: double }
        sensitivity: { type: number, format: double }
        specificity: { type: number, format: double }
        ppv: { type: number, format: double }
        npv: { type: number, format: double }
        ece: { type: number, format: double }
        brier: { type: number, format: double }
        calibrationTemperature: { type: number, format: double }
        thresholdHighSensitivity: { type: number, format: double }
        thresholdHighSpecificity: { type: number, format: double }
        stageRecommendation:
          type: string
          enum: [RESEARCH_ONLY, PILOT_DECISION_SUPPORT, CLINICAL_DECISION_SUPPORT]
        regulatoryStatus:
          type: string
          enum: [NOT_SUBMITTED, SUBMITTED, APPROVED, REJECTED]
        qaApprovedBy: { type: string }
        medicalApprovedBy: { type: string }
        reportPath: { type: string }
        notes: { type: string }
    IncidentCreateRequest:
      type: object
      required: [source, severity, title]
      properties:
        source:
          type: string
          enum: [AI_SERVICE, APP, DATABASE, PIPELINE, SECURITY, OTHER]
        severity:
          type: string
          enum: [P0, P1, P2, P3]
        title: { type: string }
        detail: { type: string }
        ownerUserId: { type: string }